requirements.txt
cairosvg==2.8.2
langchain_core==0.3.76
langchain_google_genai==2.1.12
langgraph==0.6.8
manim==0.19.0
numpy==2.3.3
Pillow==11.3.0
protobuf==6.32.1
pydantic==2.11.9
pydantic_core==2.39.0
pydantic_settings==2.11.0
python-dotenv==1.1.1
torch==2.8.0
vtracer==0.6.11
